pyproject.toml
[tool.poetry]
name = "fin"
version = "0.1.0"
description = ""
authors = ["Mohammadhossein Amouei <mohammadhossein.amouei@mail.mcgill.ca>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pyelftools = "^0.31"
tqdm = "^4.66.5"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
